refactor(tags): 純粋ルールを features/tags/domain へ集約
tag-tree.ts (hierarchy 構築) と tag-colon.ts (colon notation parse) を
features/tags/lib/ から features/tags/domain/ に move。両方とも 100%
純粋関数で DB / React / Zustand 依存ゼロ。Entry / Calendar / Review /
Chronotype と同じ domain 命名規約に揃える。

- git mv で 4 ファイル移動（履歴保持）
  - lib/tag-tree.ts → domain/tag-tree.ts
  - lib/tag-tree.test.ts → domain/__tests__/tag-tree.test.ts
  - lib/tag-colon.ts → domain/tag-colon.ts
  - lib/tag-colon.test.ts → domain/__tests__/tag-colon.test.ts
- domain/index.ts を新設し domain 内 sub-barrel を整備
- 内部 5 ファイルの deep import path を ../lib → ../domain に更新
- 既存 27 テスト（tag-tree 13 + tag-colon 14）はそのまま継続

外部 consumer は @/features/tags barrel 経由のため無変化。
仕様 / UI / DB / barrel の公開 symbol 範囲は不変。
boundary 違反なし（feature 内 directory 変更のみ）。
