+++ apps/product/src/features/tags/domain/index.ts
@@ -1,0 +1,11 @@
+export { buildTagHierarchyUpdates, buildTagTree, flattenTagTree } from './tag-tree';
+export type { TagHierarchyUpdate } from './tag-tree';
+
+export {
+  buildColonTagName,
+  getTagDisplayLabel,
+  groupTagsByPrefix,
+  hasGroupNameConflict,
+  parseColonTag,
+} from './tag-colon';
+export type { GroupedTags, ParsedColonTag } from './tag-colon';
